requirements.txt
Django==1.8.3
python-jenkins==0.4.8
